// src/type/gupy.type.ts
// Generated by https://quicktype.io

export interface GupyJobs {
  data:       GupyJob[];
  pagination: GupyPagination;
}

export interface GupyJob {
  id:                  number;
  companyId:           number;
  name:                string;
  description:         string;
  careerPageId:        number;
  careerPageName:      string;
  careerPageLogo:      string;
  type:                string;
  publishedDate:       string;
  applicationDeadline: null | string;
  isRemoteWork:        boolean;
  city:                string;
  state:               string;
  country:             string;
  jobUrl:              string;
  badges:              GupyBadges;
  disabilities:        boolean;
  workplaceType:       string;
  careerPageUrl:       string;
}

export interface GupyBadges {
  friendlyBadge: boolean;
}

export interface GupyPagination {
  offset: number;
  limit:  number;
  total:  number;
}